apps/notify-py/requirements.txt
fastapi
uvicorn
pymongo
pika
python-dotenv
